Agregando datatable server side en ventas intermediadas

+++ database/migrations/2024_06_19_194017_create_ventasintermediadas_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('nombreCliente_VentaIntermediada', 100); // Josué García Betancourt xml
             $table->dateTime('fechaHoraEmision_VentaIntermediada'); // xml
             $table->dateTime('fechaHoraCargada_VentaIntermediada')->useCurrent();
-            $table->double('montoTotal_VentaIntermediada')->unsigned(); //200.50 xml
+            $table->decimal('montoTotal_VentaIntermediada', 10, 2)->unsigned(); //200.50 xml
             $table->integer('puntosGanados_VentaIntermediada')->unsigned(); //201 (redondear el monto total del xml)
             $table->integer('puntosActuales_VentaIntermediada')->unsigned();
